Design improvement
Some web design improvement.
The footer is now okay and the small design change in terms of colors is now the good one. The colors match with Microsoft Edge in dark mode but also Mozilla Firefox and Apple Safari in dark mode.
Still not the final navigation method but still better than the old one.