Welcome to the dynamic menu
I finally arrived to implement a small menu I was already using in an other project.
I made some modifications in order to have something working normally with great color options but I still need to find a way to disable right click process on the page when the menu is open while still leaving the menu opening the page.

A little try hard for this menu but the most difficult part will be this right click...